Move UserManager from User to Manager namespace
