+++ Resources/config/services.yml
@@ -12,7 +12,7 @@ services:
             - { name: twig.extension }
 
     hackzilla_ticket.user_manager:
-        class: Hackzilla\Bundle\TicketBundle\User\UserManager
+        class: Hackzilla\Bundle\TicketBundle\Manager\UserManager
         arguments: ['@security.context', '@fos_user.user_manager']
 
     hackzilla_ticket.ticket_manager:
